Added tex source files.
I was previously keeping track of the tex documents in a
private, local repository.  I've decided to make it available
along with the pdfs.  Each directory containing pdfs, now has
a texsrc folder as well. The tex files, and images as needed, for the pdfs can be found in there.