Updated syllabus
The syllabus was updated with new dates, a new grading
scale, spruced up expectations and collaboration policies,
and assignments list. The biggest change is the fact that I
dropped the textbook. I wasn't using it enough to justify
the cost to the students.  I also had to do too much dancing
to get it to mesh with the course.