Week 6 complete. Resolves #9.
The rewrite of LN7-9 is complete with this commit. The new
path to effect based procedures is through a simple interactive
testing problem. Polymorphism is introduced as a means of
testing I/O procedures meant for cin/cout using string streams.
Pass by reference and returning references for chained
effects is also discussed and demonstrated.