search/sort/proj2 reorg
Old 14 (profiling with chrono) is now 16. The 15+16 combo sets us
up for theoretical and actual performance (time) analysis of
procedures. The new 17 (work in progress here) goes through the
analysis of search and sort from 14. THe new 18 deals with
randomized procedures in preparation for the project, which
requires vectors of randomly shuffled data. The new 19 goes over
the Mathematica required for project 2 (basic list management,
tables, point plots, and import and export).